// src/app/models/BuscarRegistro.ts
export class BuscarRegistro {
    nmbproducto1: string;
	nroregsan: string;
	fchvencimiento: string;
	nombre: string;
    modcompuesta: string;
    des_estadoregistro: string;
    nbr_grupo: string;
    nroexpediente: string;
    cdgproducto: string;
    cdgexpediente: string;
    TITULAR: string;
    principio_Activo: string;        
    concetracion: string;
}