// src/app/models/TrReportedispcomercializacion.ts
export class TrReportedispcomercializacion {
    idreportedispcomercializacion: number;
	numeroexpediente: string;
	numeroregistrosanitario: string;
	fechainivigenciadisp: Date;
	nombremodalidad: string;
	razonsocialtitular: string;
	nombreproducto: string;
	nombreprincipioactivo: string;
	concentracionppioactivo: string;
	codreportedisponibilidad: number;
	motivoreportedisp: string;
	fechafinvigenciadisp: Date;
	indnotificacioncorreo: string;
	emailnotificacion: string;
	nombrearchivo: string;
	ruta: string;
	estado: string;
	fechareportedisp: Date;
	idusuario: number;
	idestablecimiento: number;
}